+++ frontend/app/config.py
@@ -2,7 +2,7 @@
 import os
 
 # Setup api url
-BONSAI_API_URL = os.getenv("BONSAI_API_URL", "http://localhost:8011")
+BONSAI_API_URL = os.getenv("BONSAI_API_URL", "http://api:8000")
 # where reference genomes are found
 DATA_DIR = os.getenv("DATA_DIR", "/tmp/data")